frontEnd/requirements.txt
requests
PyJWT
reportlab
